Moved methods into `PaymentQueueWrapper`
