+++ Sources/Purchasing/Purchases/Purchases.swift
@@ -249,6 +249,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
     private let receiptFetcher: ReceiptFetcher
     private let requestFetcher: StoreKitRequestFetcher
     private let storeKitWrapper: StoreKitWrapper?
+    private let paymentQueueWrapper: PaymentQueueWrapper
     private let systemInfo: SystemInfo
     private var customerInfoObservationDisposable: (() -> Void)?
 
@@ -290,6 +291,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
         let storeKitWrapper: StoreKitWrapper? = systemInfo.storeKit2Setting.shouldOnlyUseStoreKit2
         ? nil
         : StoreKitWrapper()
+        let paymentQueueWrapper = storeKitWrapper?.createPaymentQueueWrapper() ?? .init()
 
         let offeringsFactory = OfferingsFactory()
         let userDefaults = userDefaults ?? UserDefaults.standard
@@ -391,6 +393,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
                   attributionPoster: attributionPoster,
                   backend: backend,
                   storeKitWrapper: storeKitWrapper,
+                  paymentQueueWrapper: paymentQueueWrapper,
                   notificationCenter: NotificationCenter.default,
                   systemInfo: systemInfo,
                   offeringsFactory: offeringsFactory,
@@ -413,6 +416,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
          attributionPoster: AttributionPoster,
          backend: Backend,
          storeKitWrapper: StoreKitWrapper?,
+         paymentQueueWrapper: PaymentQueueWrapper,
          notificationCenter: NotificationCenter,
          systemInfo: SystemInfo,
          offeringsFactory: OfferingsFactory,
@@ -441,6 +445,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
         self.attributionPoster = attributionPoster
         self.backend = backend
         self.storeKitWrapper = storeKitWrapper
+        self.paymentQueueWrapper = paymentQueueWrapper
         self.offeringsFactory = offeringsFactory
         self.deviceCache = deviceCache
         self.identityManager = identityManager
@@ -1223,8 +1228,7 @@ public extension Purchases {
      */
     @available(iOS 13.4, macCatalyst 13.4, *)
     @objc func showPriceConsentIfNeeded() {
-        // TODO: SK2?
-        self.storeKitWrapper?.showPriceConsentIfNeeded()
+        self.paymentQueueWrapper.showPriceConsentIfNeeded()
     }
 #endif
 
@@ -1260,8 +1264,7 @@ public extension Purchases {
     @available(macOS, unavailable)
     @available(macCatalyst, unavailable)
     @objc func presentCodeRedemptionSheet() {
-        // TODO: SK2?
-        self.storeKitWrapper?.presentCodeRedemptionSheet()
+        self.paymentQueueWrapper.presentCodeRedemptionSheet()
     }
 #endif
