+++ Sources/Purchasing/StoreKit1/StoreKitWrapper.swift
@@ -87,24 +87,6 @@ class StoreKitWrapper: NSObject, SKPaymentTransactionObserver {
         return SKPaymentQueue.canMakePayments()
     }
 
-    @available(iOS 14.0, *)
-    @available(macOS, unavailable)
-    @available(tvOS, unavailable)
-    @available(watchOS, unavailable)
-    @available(macCatalyst, unavailable)
-    func presentCodeRedemptionSheet() {
-        // Even though the docs in `SKPaymentQueue.presentCodeRedemptionSheet`
-        // say that it's available on Catalyst 14.0, there is a note:
-        // This function doesn’t affect Mac apps built with Mac Catalyst.
-        // It crashes when called both from Catalyst and also when running as "Designed for iPad".
-        if self.paymentQueue.responds(to: #selector(SKPaymentQueue.presentCodeRedemptionSheet)) {
-            Logger.debug(Strings.purchase.presenting_code_redemption_sheet)
-            self.paymentQueue.presentCodeRedemptionSheet()
-        } else {
-            Logger.appleError(Strings.purchase.unable_to_present_redemption_sheet)
-        }
-    }
-
     func payment(withProduct product: SK1Product) -> SKMutablePayment {
         let payment = SKMutablePayment(product: product)
 
@@ -120,13 +102,6 @@ class StoreKitWrapper: NSObject, SKPaymentTransactionObserver {
         payment.paymentDiscount = discount
         return payment
     }
-
-#if os(iOS) || targetEnvironment(macCatalyst)
-    @available(iOS 13.4, macCatalyst 13.4, *)
-    func showPriceConsentIfNeeded() {
-        self.paymentQueue.showPriceConsentIfNeeded()
-    }
-#endif
 
 }
 
@@ -177,6 +152,15 @@ extension StoreKitWrapper: SKPaymentQueueDelegate {
     // Sent when the storefront for the payment queue has changed.
     func paymentQueueDidChangeStorefront(_ queue: SKPaymentQueue) {
         self.delegate?.storeKitWrapperDidChangeStorefront(self)
+    }
+
+}
+
+extension StoreKitWrapper {
+
+    /// Creates a `PaymentQueueWrapper` backed by the same `SKPaymentQueue`.
+    func createPaymentQueueWrapper() -> PaymentQueueWrapper {
+        return .init(paymentQueue: self.paymentQueue)
     }
 
 }
